features/gateway/PMOVES-BotZ-gateway/mcp-proxy-server/requirements.txt
fastmcp==2.13.0
validators==0.35.0
PyYAML==6.0.2
starlette==0.40.0
uvicorn>=0.31.1
prometheus-client>=0.20.0

